pyproject.toml
[project]
name = "crepetrt"
version = "0.1.0"
description = "Fast trt inference for onnx CREPE models"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "joblib>=1.5.1",
    "librosa>=0.11.0",
    "matplotlib>=3.10.5",
    "numpy>=2.2.6",
    "onnx>=1.19.0",
    "onnxruntime-gpu>=1.22.0",
    "pandas>=2.3.2",
    "penn>=1.0.0",
    "tensorrt>=10.13.2.6",
    "tensorrt-cu12>=10.13.2.6",
    "tensorrt-dispatch-cu12>=10.13.2.6",
    "tensorrt-lean-cu12>=10.13.2.6",
    "torbi",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
    "tqdm>=4.67.1",
    "typer>=0.16.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-multipart>=0.0.17",
    "coloredlogs>=15.0.1",
    "soundfile>=0.13.1",
]
[project.scripts]
crepetrt = "crepetrt.inference:app"
crepetrt-serve = "crepetrt.inference:serve"
crepetrt-predict = "crepetrt.inference:predict"

[tool.setuptools.packages.find]
where = ["."]
include = ["crepetrt*"]

[tool.uv.sources]
torbi = { git = "https://github.com/korguchi/torbi.git" }

[dependency-groups]
benchmark = ["torchcrepe>=0.0.24"]
dev = [
    "line-profiler>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
dev = [
    "ipykernel>=6.30.1",
    "line-profiler>=5.0.0",
    "mir-eval>=0.8.2",
    "pytest>=8.4.1",
]
